// web/apps/webapp/src/queries/__generated__/organizationTeamsTab_teams_query.graphql.ts
export type organizationTeamsTab_teams_query$data = {
  readonly teams: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly name: string;
        readonly organization: {
          readonly name: string;
          readonly uniqueId: string;
        } | null | undefined;
      };
    }>;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "organizationTeamsTab_teams_query";
};
export type organizationTeamsTab_teams_query$key = {
  readonly " $data"?: organizationTeamsTab_teams_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"organizationTeamsTab_teams_query">;
};